move message to fragment module
