+++ chain-impl-mockchain/src/block/mod.rs
@@ -1,6 +1,6 @@
 //! Representation of the block in the mockchain.
 use crate::key::Hash;
-use crate::message::{Fragment, FragmentRaw};
+use crate::fragment::{Fragment, FragmentRaw};
 use chain_core::mempack::{read_from_raw, ReadBuf, ReadError, Readable};
 use chain_core::property::{self, Serialize};
