+++ chain-impl-mockchain/src/testing/ledger.rs
@@ -3,8 +3,8 @@ use crate::block::HeaderHash;
 use crate::config::ConfigParam;
 use crate::ledger::Error;
 use crate::ledger::Ledger;
-use crate::message::config::ConfigParams;
-use crate::message::Fragment;
+use crate::fragment::config::ConfigParams;
+use crate::fragment::Fragment;
 use crate::milli::Milli;
 use crate::transaction::*;
 use chain_addr::{Address, Discrimination};
